Extract tab container to own component
