+++ src/components/tabs/TabItemsContainer.vue
@@ -1,0 +1,5 @@
+<template>
+    <div class="flex items-center flex-col sm:flex-row">
+        <slot></slot>
+    </div>
+</template>
